+++ practice-app/package.json
@@ -1,5 +1,6 @@
 {
   "name": "practice-app",
+  "description": "Sample react app to practice different designs, tools and concepts",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
